requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.2.0
blinker==1.8.2
Bootstrap-Flask==2.3.3
buildozer==1.5.0
cachetools==5.5.2
canvas==0.35.0
category-encoders==2.6.3
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colorgram.py==1.2.0
comm==0.2.2
comtypes==1.2.0
configparser==7.2.0
contourpy==1.2.1
cookiecutter==2.6.0
cron-converter==1.2.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.9
Django==5.2.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-utils==0.15.0
dnspython==2.6.1
docutils==0.21.2
docx==0.2.4
dominate==2.9.1
EasyProcess==1.1
ecapture==2.0.2
email_validator==2.2.0
entrypoint2==1.1
env-tools==2.4.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
Flask==2.3.2
flask-bootstrap5==0.1.dev1
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.4.1
google-auth==2.40.3
google-genai==1.23.0
greenlet==3.0.3
grpcio==1.71.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.28.1
idna==3.4
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.3.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.0.0
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.5
mypy==1.15.0
mypy_extensions==1.1.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.25.1
opencv-python==4.8.0.74
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pefile==2023.2.7
pexpect==4.9.0
phonenumbers==8.13.18
Pillow==10.0.0
platformdirs==4.3.7
plotly==6.1.2
plyer==2.1.0
prettytable==3.10.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.5.0
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pypdf==3.15.0
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-pptx==1.0.2
python-slugify==8.0.4
python-telegram-bot==21.4
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
redis==5.0.7
referencing==0.36.2
regex==2024.7.24
reportlab==4.4.0
requests==2.31.0
RestrictedPython==8.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.22.0
Send2Trash==1.8.3
sentry-sdk==2.27.0
sh==2.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
spotipy==2.24.0
SQLAlchemy==2.0.25
sqlparse==0.5.3
stack-data==0.6.3
statsd==4.0.1
statsmodels==0.14.2
telegram==0.0.1
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.2.0
tini==4.0.0
tinycss2==1.4.0
tk==0.1.0
tkcalendar==1.6.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
twilio==9.2.1
typer==0.15.3
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20250402
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.31.2
visitor==0.1.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.0
widgetsnbextension==4.0.14
wikipedia==1.4.0
win10toast==0.9
wolframalpha==5.0.0
wsproto==1.2.0
WTForms==3.0.1
XlsxWriter==3.2.3
xmltodict==0.13.0
yarl==1.9.4
zipp==3.21.0
